Archive/requirements.txt
asyncio==3.4.3
callofduty.py==1.2.2
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
h11==0.11.0
httpcore==0.12.2
httpx==0.16.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
rfc3986==1.4.0
sniffio==1.2.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1